Interpret more application states
Currently only STARTED is recognized and interpreted as module state
`deployed`. This commit recognizes all the application instance states
we are aware of, and interprets them as various module instance states.

It also attempts to obtain application instance status irrespective of
the reported state of the application, and ignores the exception that
might be thrown.