Destroy Failed Deploys
If a deployment fails on CloudFoundry the resulting apps may be in the
"DOWN" state. This state was not allowed for in the calculation of
module status, which meant that the module was classed as 'unknown'.
Modules in that status were not undeployed, hence the apps were orphaned
on CF. This change adds the missing state, and hence allows such modules
to be deleted as expected.