Add application environment variable to 'mark' module applications.
Previously a fixed application name prefix was used to identify module
applications and the moduleId was deduced from the name. This commit adds an
environment variable to the application which contains the string version of
the ModuleId. The prefixing is retained, as a means of recognising module
applications and their routes from the CF command line, but only the embedded
moduleId is used to identify module applications.