Align dependencies with BOMs in Spring Cloud
To allow users to manage dependencies in application projects
for data flow and the rest of Spring Cloud independently, we
need to clean up the BOM a bit and use the Cloud BOMs individually
in the libraries in data flow.

Also fixes the wrapper configuration so Maven can be used on the
command line to do a manual deployment.