Add additional env vars to deployed module
User can specify additional command line args like this in the
local deployer app:

deployer:
  local:
    sharedArgs:
      includes: org.springframework.cloud:spring-cloud-starter-config:1.1.0.BUILD-SNAPSHOT
      spring.application.name: ${module.label}
      spring.cloud.config.uri: http://localhost:8888

Then all modules are launched with "includes" on the classpath
and enough information to bootstrap the config client.

The "includes" is a feature of the module launcher, and it is a
comma separated list, so we have to be careful to merge
module and deployment specific arguments with the same name.