Introducing composed runner
= Composed Task Spike

The purpose of this spike is to show how the Composed Tasks will be implemented in
the Data Flow project.

== Design Document
If you hadn't had a chance please review the design document.  It can be found:
https://docs.google.com/document/d/10FiyNrKH1KmfDj7WRVmREc9VwzSZx3XnMBP8VU8y99U/edit#heading=h.4h9zsajkx156[here]

=== What is included

Based on the design document this story created a spring-cloud-dataflow-composed-task sub-project
under the Spring Cloud Data Flow Project.  In this project you will see that the TaskLaunchTasklet, ComposedTaskRunner
along with the supporting ComposedTaskConfiguration and ComposedTaskStepExecutionListener are included.
Also you will note that the DataFlowServerConfiguration has been updated to include a EventListener to register the ComposedTaskRunner as a task in the repository (as specified in the design document) at Data Flow Server startup.  Thus you will see that a `composed-task-runner` task appears upon executing a `app list`.  This registration can be disabled by setting the spring.cloud.dataflow.composed.task.enabled property to false.

=== What is not included

In this spike the following has not be included:

* Controllers for task compose
* Parser for the DSL within the ComposedTaskRunner (you can only include a single task definition for the DSL at the moment).
* Parent Task / Child Task relationship.

==== Kicking the tires

`task create mytask --definition "composed-task-runner --composedTaskDSL='AAA'"`  Where *AAA* is your task definition.
`task launch mytask`

resolves #1085