Introduce parser for composed task DSL parsing
This commit introduces the composed task DSL parser.

Rather than duplicate the existing tokenizer for stream,
that has been abstracted into a common tokenizer base
class that both the stream and task tokenizers extend.

There is a new ComposedTaskParser that produces an AST
for composed task DSL strings.

A visitor ComposedTaskVisitor is provided for any code
needing to walk the AST. There are two provided visitors
in this commit - the ComposedTaskValidatorVisitor
verifies the AST (since it is possible to have text
that follows the correct syntax but is still invalid)
whilst the GraphGeneratorVisitor is included here but
won't be used in anger until the client visualization
of composed tasks needs it.