gh-ui-158 Add property `formLogin` to `SecurityInfoResource`
Expand the `SecurityInfoResource` by a property `formLogin` that indicates that formLogin is being used. This is needed in cases where the application needs to be "reccreated" using the xauth-token and the `SecurityInfoResource` (aka the user presses refresh) only. If Outh2 authentication is being used, this property will be `false`.

Is depended on by https://github.com/spring-cloud/spring-cloud-dataflow-ui/issues/158