Adjust text-graph and graph-text algorithm and improve validation
The adjustments are made to account for spaces. If a graph node property
(which will manifest as an app argument) has a space in it, and isn't
already quoted, then quotes will be added in graph-to-text. Also, exit
status names will always get quoted in graph-to-text.  In a similar
manner, text-to-graph is adjusted to remove the quotes. This means
we don't need to muck about on the client side forcing users to add
quotes.

Validation will now report a problem if the name of an app
used in a transition clashes with another app in the DSL string
and a label should be used.

Also, a number of tests have been removed which relate for more
advanced DSL syntax. Rather than rework the tests now they will be
reactivated when we decide what the final form of the DSL
actually is.

Add extra test

Alter name generation for composed task to use taskname-appname format

Use TaskNode.getTaskPrefix method for childTaskPrefix in deleteTaskDefinition