Support newline Characters in Quoted Strings
Certain properties are mapped to `Properties` objects using Spring's `PropertiesEditor`.

The sytax for such properties is `someProp='foo=bar \n baz=qux'`.

Currently the abstract tokenizer terminates a quoted token when `\n` or `\r` is found.

Change the `isArgValueIdentifierTerminator` to only terminate on such characters if we are
not parsing a quoted token.