Add infrastructure for documenting OOTB apps in SCDF
Adds a adoc file that grabs the READMEs of OOTB apps, similar to what is done
in the starters reference doc. Also includes an SCDF specific file per app
that is meant to document the usage of the app in SCDF (eg with DSL as part
of a stream).

The file is checked for basic syntax, and supporting files are created empty
via a unit Test.

Fixes #1574
Fixes #1575
Fixes #1576