Add checksum feature to about controller
also added base set of tests.

resolves #1779

resolves #1758

This PR allows the user to configure a link to the downloadable jars that comprise SCDF.
From these links the about will also retrieve the sha1 checksum and or the sha256 checksum
If the checksum is not available it will not display the value as a part of the results.