Change from tocify to tocbot
Spring's other projects are headed in the direction of using Tocbot, which improves on tocify by needing fewer libraries to work. (In fact, it has no dependencies.) It also works with our other features, such as content toggling.

Fixed a bug to restor the ToC title

"Table of Contents" went missing. I restore it. Credit to Gunnar Hillert for finding the flaw.`