New content to address Spring Cloud Dataflow #2042
Link to the issue: https://github.com/spring-cloud/spring-cloud-dataflow/issues/2042

I added a chapter for a new Task Developer Guide. That required the addition of seven new images, too. I also caught a couple of minor issues in the Streams Developer Guide while I was using it as a model for the new file.

BIG thanks to Glenn Renfro (cppfws on Github) for answering a bunch of questions and for updating the demo while I was writing about it (to make it work a little better in conjunction with the guide).