Improve composed task graph to text algorithm for nested splits
A nested split is when a split includes a further split as one of
its branches. Without these changes 3 links from a node would
simply be considered as a 3 way split, but in fact it can be
a 2 way split with a nested split. For example: <<AA || BB> && CC || DD>

Fixes #712