Adjust from comma to double pipe for separating unbound stream apps
This adjusts the syntax so that between unbounded stream apps
you need to use ||, for example: aaa || bbb
This is contrast to regular stream apps that continue to use
single pipe: aaa | bbb
You cannot use channels with unbounded stream apps and all the tests
and errors messages related to the previous comma syntax have been
updated to reflect the new syntax.

I have also merged in the PR I created that handles content
assist for unbounded stream apps and adjusted that to use the new
syntax.