Optimize the performance of job step count query
resolves #2583

The purpose of this PR is to optimize query required to populate the job executions for the UI and shell.

In both cases they require only the number of step executions.   However the current query returns a fully qualified JobExecution object which contains all the step executions associated with the step execution.  The problem is that if a user has a large number of steps with the JobExecution, the result takes a long time to return.   Up to 60 seconds ffor 100 JobExecutions with 500  step executions per.

To resolve this issue a new endpoint was created job/executions/jobexecutioninfoonly that returns a JobResource that contains a JobExecution without its associated stepexecutions, however the stepExecutionCount is populated in the resource.

The new endpoint was added so that if others were currently using the fully qualified jobs from the /job/executions endpoint, they would not be impacted.