Produce error if single char stream element arg value is special char
Without this change the system was not consistent in how it treated
special characters in argument values (e.g. | or >). They were
not treated as special if they were the first character in the
argument value (e.g. 'app --param=|'). When used like that they
should still be treated as special (i.e. end of stream def) and
so an error should come out about a missing argument value. If
they are truly the argument value they should be quoted.

Resolves #2499