Refactor DefaultTaskExecutionService's constructor
resolves #2768

When creating the TaskAppDeploymentRequestCreator I couldn't see the need for creating a service with @Transactional.
It could be referring to the creation of the TaskExecution that does require a new transaction, however this transaction information
is required by the Auditing and for post task launch request processing.  Similarly TaskExecutionInformation is required
for both request creation and auditing.