Allow splits with a single entry for some composed task scenarios
With this change the DSL parser allows a single entry in a split - this
is so that you can have a task that uses transitions and rather than
those transitions jumping straight to the END after they are done, they
will jump to the task following the split. Effectively the split is
being used to create a nested flow. This is not a new usage of it, but
previously it only allowed nested flows involving multiple tasks to
run in parallel. Now you can use it for just a single task in a
nested flow.

The parser has been changed and the graph-to-text and text-to-graph
algorithms have been updated to support it.

Resolves #3263