Remove checks for missing steps in step execution list code.
resolves #3381

In short the code that was removed made sure that stepExecutions that are associated with a job returned from the jobLocator are represented in the step execution list.
However, the map within the jobLocator is never populated by dataflow.  In fact jobLocator's usage in this class is really a no-op.

The cause of the problem the user reported came from the fact that roughly the same query was
executed 2 times within the same method and from time to time the batch app would update the batch_step_execution table with a new entry,
and thus the dataflow code would interpret this as a stepExecution that was in the jobLocator job and thus enter stubbed entry into the resultset.
The fix was to remove that extraneous code.