Add Scheduling to CI/CD Task Effort
Note:
1) I will open a issus efor the UI and for core:
Currently if a user try to delete all schedules associated with a task definition it will fail to work.   This is because we no longer use just the app name.  It is now scdf-appname-schedulename
However if you delete the task definition it will delete all associated schedules (as it should).

2) You will see in the the application.yml and in TaskConfigurationProperties the maven URI for the scheduler-task-launcher is pointing to a snapshot.  Is there a way we can automate this value vs. having to do it via the release process?