Removed the type hint from the job params on job restarts.
resolved #4204

The original intent was to use the entire key that was returned from batch.
However when relaunching a job from CF it will fail.  That is because the build pack requires a triple escape to use `(` character ie.`\\\(`
This also causes a problem when passing this value as a property when using shell entrypoint for k8s.

So since this is just a hint to batch, we will remove this info and let batch figure out the correct type.