CTR will not attempt to connect to SCDF until it is required.
This allows CTR to record its information to task execution table in the event that the host port is incorrect

Moved Relaxed names so that it can be shared

Removed the scdf-core-server dependency.

Not sure how to remove the Relaxed names since the registrar creates its own BeanDefinitionBuilder using the ComposedTaskRunnerStepFactory class.
And thus needs to be given the properties at construction time.

Removed dependency version from CTR POM and using parent pom version