Complete task task resource cleanup on task deletion.
If there is no task deployment when a task definition is deleted SCDF will delete all app resources for the task definition name across all platforms.
If there is a task deployment for that task definition, then it will only delete those app resources associated with that task deployment.
This is to resolve the case where a task launch failed on the platform and thus no task deployment record is created.  Hence we don't know what platform the launch attempt occured.

resolves #4574