Launching the same CTR instance will now succeed (#5415)
**Scenario**
CTR launch with simple conditional statement that contains both Boot2 and Boot3 apps.  Relaunch the CTR again it should succeed by default.

```
app register --bootVersion 3 --name timestamp3 --uri maven://io.spring:timestamp-task:3.0.0 --type task
task create --name ctrseq --definition "timestamp && timestamp3"
task launch --name ctrseq
task list
task execution list
task execution status --id ctrseq
task launch --name ctrseq
task execution list
task execution status --id ctrseq
```
The cause is SCDF stores the taskPrefix using both the composed-task-app-properties and the properties.   While this works for the first execution,
The second execution fails because it picks up a duplicate property for the taskPrefix.   This is because on the first execution, it is stored in the manifest as a base64 value.
So the solution is not to re-add the taskPrefix app properties and use just the composed-task-app-properties, which is used by CTR to determine which prefix to use and will be passed to the child app.