Await for all audits to be recorded
AuditRecordControllerTests ocassionally fails on the
testRetrieveAuditRecordsFromGivenDateToNull test.

This is because it is expecting 2 audits but ocassionally gets 3
because a create audit for the last entry isn't recorded till
after the endTime is recorded.

This commit adds a 30 second max wait for all create audits to complete.

In this case add