Script should not end terminal if not using bash
If the user hit the exit 0 (when using source) and
they did not use bash  terminal would exit.
The user would not know why the terminal exited.

Updated script to just check to see if there is bash available.

Set namespace to default if not specified

Add additional instructions when not running in bash

Document the use of the --namespace key

namespace is no longer positional but requires a --namespace or -ns key