What was done For each database type the following migration was implemented:
Prefix the TASK V2 and BATCH V4 tables with V2_ .
This allows user to determine what they wish to do with this data
Remove BOOT3_prefix for TASK V2 and BATCH V5 tables
Make sure that these migrations were supported by flyway
Make sure that the Migration SQL scripts were added to associated yaml files.
What was affected
The following databases were migrated:

H2
MariaDB
Mysql
Postgres
Oracle
SQLServer
DB2
Types of migrations
There were 3 types of migrations that occurred.

Default - These types of migrations were typically alter tables and alter sequences.
The following databases belong to this group:
a. MariaDB
b. Mysql
c. Postgres
d. Oracle (with some exceptions)
e. Sql Server (commands different than others, but principles remained).

In-Memory - For this case the Task V2 and Batch V4 tables/sequences DDL
was removed and BOOT3_ prefix was removed. The only in-memory we support is H2.

DB2 - Gets its own category
To rename the tables db2 requires all primary and foreign keys to be dropped.
In this case we decided to create the new tables and copy the contents of the
original tables to the new tables. Then remove the original.
This was to avoid errors when recreating the keys

How to eat this Elephant
This PR is a bit large so let's discuss how we can handle this review.
Let's review this by sampling one database from each type of migration.
This is so that we can make sure the general pattern works for folks.

Default type, look at Mariadb
In Memory type look at H2
DB2 Type look at (well... ) DB2 :-D
After we finish this sampling review, those changes will be applied to the other databases and then a full review
can be made.