Task and Batch tables should use default prefixes
Remove BOOT3_prefix for batch 5 and task 2 tables for mariadb

Add V2_ prefix for batch 4 and task 2 tables

H2 and Postgres updates

* H2 Removed V2 Task and V4 Batch tables and removed BOOT3 prefix.
  * Removed table renames for V2 task tables

* Postgres added migration script removing BOOT3_ prefix for Batch V5 and Task V3

Postgres and Mariadb work

Remove BOOT3_prefix for batch 5 and task 2 tables for sqlserver

Migrate Task and Batch tables to Boot3 for Oracle

Migrate SCDF flyway db2 schema

To rename the tables db2 requires all primary and foreign keys to be dropped.
In this case we decided to create the new tables and copy the contents of the original tables to the new tables.  Then remove the original.
This was to avoid errors when readding the keys

Create mysql flyway migration scripts

Update applicaiton-init yaml files with the new migration scripts

Re-add test scope