Add Autoconfiguration.imports to resource includes in pom.xmls
We had an issue where some of the auto configurations were not firing.
This is because when a user specifies the resource tag in maven,
the boot plugin expects the user to fill the resources manually vs. Its default behavior

Also removed the version from the H2 dependency so that it can be managed by the bom

Removed unnecessary @Configuration annotaions where a @Autoconfiguration annotation is present