Update Code to use JobRepository bean directly
Since we will only support BOOT 3 we do not need the JobRepositoryContainer
to retrieve BOOT3 or BOOT 2 based JobRepositories.

Update test code to set the default time for the local date time sample to
a default of midnight

Allow services to use JobServie and JobExplorer directly

Currently we use containers to allocate the JobService and JobExplorer based on boot version.

This is no longer necessary.  So this PR removes these containers

Update Project as to restore tests success percentage to original state

After the updates below tests that were passing started to fail.
Some because of Batch 5 updates, but others because of the removal
of some of the container classes.

Update code based on code review comments

Removed todo
Removed ExtendsWith statement