Reuse CriteriaQuery when retrieving result count.
When migrating to the latest Hibernate we saw the following exception:
```
Caused by: java.lang.IllegalArgumentException: Already registered a copy: SqmBasicValuedSimplePath
```
This is because we were recreating a CriteriaQuery for our Queries.
Hibernate no longer allows us to do that, but rather allows us to use the existing
CriteriaQuery, but use the convenience method createCountQuery to
provide the criteria query for our createQuery.

Also removed 6.1.7 versionfor hibernate core  that allows us to use the Boot Bom.
This also caused some downstrem issues where dataflow was using the old version brought in from skipper