Task services should not require the uses of Schema version.
Phase 1 migration of removing Aggregates.
The purpose of this PR is to allow users to execute and query results of  Boot3 Tasks only.
While there are some Task services that still use aggregate tables, it is because
they are associated with Jobs which will be handled in Phase 2

This is phase 1 of removing SchemaVersion from SCDF.
This commitss starts the work in the TaskExecutionController.
This finishes teh conversion in DefaultTaskDeleteService