User needs ability to specify app version when creating schedule
This update allows user to specify version.label=<version.number>
Tests were also updated because the original settings assumed that the
 appregistry was real instance instead of being mocked. Thus the find would always return null.
And in this case the tests returned a false positive.
Now that the mocks are in place it excercises all the code.
Also added explicit test if user does not set the version number.  Some of the tests do this,
but wanted an explicit test to verify this.

resolves #5705