Need to select the serialization method for JobParameters for the commandline
User needs ability to set the default serialization technique for SCDF when restarting a job
User needs ability to select a serialization technique for a specific job restart
When user restarts a job repository from the list that is derived from thinjobexecutions it should use default technique

Add restful documentation.

Add support to allow user to set useJsonJobParameters for job relaunch via the shell.

Note: there are not tests for the shell update in this commit. This is because the current
set of tests rely on @EnableDataflowServer which does not work.   But before we fix
@EnableDataflowServer we need to make sure we want to carry it forward per Issue #1040

Polish PR before push to repo

JobCommand should ignore identifyingParameters when deserializing JobParameters

This is a generated list and and will cause deserialization to fail if not skipped

SCDF needs to support any type of class for JobParameters

However, if the class is not a base java type or one provided by dataflow the user has to build dataflow with their class included.

Add tests for JobParameterJacksonDeserializer

Remove Disabled annotation from JobExecutionController tests that are no longer needed

Update per code review request.

Added test for JobParamterMixin via the JobTemplateTests
Removed the duration parameter from getDurationBasedEstimate method
Rebased

Reset the duration calculation from nanos back to millis.

Optimized restartExecutionArgs routine that removes duplicates.

This was per a comment in code review

Remove unnecessary exclusions from AbstractShellIntegrationTest

The changes are  code review requests.

Add warn log message when job restart id is invalid.

* Initialize ObjectMapper with the tools provided by DataFlowTemplate when testing

* These changes were identified during code review