Remove Deprecated methods from TaskConfigurationProperties.
This will require updates in other classes to have them use ComposedTaskRunnerConfigurationProperties
Remove the the utilities methods from TaskServiceUtils

Add Tests for ComposedTaskRunnerConfigurationProperties

This test suite was originally located in the TaskServiceUtils.  But upon the removal of the Deprecation these tests were removed from its suite of tests.
These tests are still valid were migrated to a suite for the ComposedTaskRunnerConfigurationProperties.