Remove Deprecations or Deprecated methods from JobService
In 2.11.x the methods in JobService were marked as Deprecated as
Dataflow was using the Aggregate infrastructure.

In 3.x SCDF is dropping the Aggregate infrastructure and will use JobService instead.
But 2 cases arose during the removal of the Deprecated annotation.
While methods that are used after the 3.0 migration, the Deprecated annotation was removed.

Signed-off-by: Glenn Renfro <grenfro@vmware.com>

Update what's new doc with deprecation removals