Remove deprecations from AppRegistryController
The original issue states to remove deprecated functions.
However, this would remove some of SCDF's default behavior when registering applications.
In that it can look the suffix of the URL to obtain the version.
The original Issues for the PR that put in the deprecations just wanted to give the
user the ability to override the version.

Also deprecated the use of bootVersion when registering an app.

Resolves #6011

Signed-off-by: Glenn Renfro <grenfro@vmware.com>